prediction_ms/requirements.txt
Django==3.2.12
djangorestframework==3.14.0
djongo==1.3.6
pandas== 1.5.1
numpy==1.23.3
keras==2.13.1
pandas-datareader==0.10.0
yfinance==0.2.28
scikit-learn==1.3.0
tensorflow==2.13.0
yahoo_fin==0.8.9.1
pymongo==3.12.3
django-crontab==0.7.1
django-cors-headers
